Revert commits
Bootstrap is being relicensed from Apache 2.0 to MIT so that more of our community can use the free and open source software we lovely create together. As commits are copyright their original authors, the commits of any individual not agreeing to relicense their contributions are being removed to be later rewritten by the @twbs/team.

This undoes the changes introduced by the following commits:
https://github.com/twbs/bootstrap/commit/66051e06bafce0d730af102b939b4b5bae3232c3 , https://github.com/twbs/bootstrap/commit/6dc895403af70937e0ea47a1d7518b5628aeac02 , https://github.com/twbs/bootstrap/commit/4c414219d22056a10bfd47766b1a193db9c4ae60 , https://github.com/twbs/bootstrap/commit/eadd557eea07a4557ec5de3c86730bb7584eacb3 , https://github.com/twbs/bootstrap/commit/eb416bc51bf6be1f85454cf3561d544cbabeedfd , https://github.com/twbs/bootstrap/commit/e05ec80ff912d558aada9a9bc3092f0d228f1ecc , https://github.com/twbs/bootstrap/commit/4756978bae47ebfdc60e06509e5bd6784b2b90f3