Increase contrast in dark-mode (#38525)
This increases the default brightness for the body text-color in dark
mode, which inceases readability a bit.

It was also suggested to darken the background a bit, but we
unfortunately don't have a darker gray in our color palette. So for now,
let's increase the text-color by two notches, and leave the background
color as-is instead.